# BankingSystemAPIs/apidocs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Banking System API Documentation",
        "description": "The Banking System API provides a comprehensive set of endpoints for managing banking-related operations. It is designed to support various banking activities such as account management, transaction processing, and exchange rate management. This API is built with scalability and security in mind, ensuring seamless integration with client applications.\r\n\r\nKey Features:\r\nAccount Management: Create, retrieve, update, and delete accounts. Perform operations such as deposits, withdrawals, and transfers.\r\nTransaction Processing: Retrieve transaction histories, manage transaction records, and delete transactions securely.\r\nExchange Rates: Manage exchange rates for supported currencies, including retrieval, creation, updates, and deletions.\r\nLogging: Detailed request and response logging to facilitate debugging and auditing.\r\nRESTful Design: Follows RESTful principles with predictable endpoints, standard HTTP methods, and meaningful status codes.\r\nSwagger Documentation: Integrated Swagger/OpenAPI support for testing and visualizing API capabilities.",
        "contact": {
            "name": "Mahmoud Metwalli",
            "email": "mammkbih@gmail.com"
        },
        "version": "v1"
    },
    "paths": {
...
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Retrieves a client by ID.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The unique identifier of the client.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Client"
                ],
                "summary": "Updates an existing client by ID.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The unique identifier of the client.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "The updated client details.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Client"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Client"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/Client"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Client"
                ],
                "summary": "Deletes a client by ID.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The unique identifier of the client.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Retrieves all clients.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Client"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Client"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Client"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Client"
                ],
                "summary": "Creates a new client.",
                "requestBody": {
                    "description": "The client details.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Account": {
                "required": [
                    "accountNumber",
                    "balance",
                    "clientId",
                    "currency"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "accountNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "balance": {
                        "type": "number",
                        "format": "double"
                    },
                    "rowVersion": {
                        "type": "string",
                        "format": "byte",
                        "nullable": true
                    },
                    "currency": {
                        "maxLength": 3,
                        "minLength": 1,
                        "type": "string"
                    },
                    "rate": {
                        "$ref": "#/components/schemas/Rate"
                    },
                    "clientId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "accountsTransactions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AccountTransaction"
                        },
                        "nullable": true,
                        "readOnly": true
                    },
                    "client": {
                        "$ref": "#/components/schemas/Client"
                    }
                },
                "additionalProperties": false
            },
            "AccountTransaction": {
                "type": "object",
                "properties": {
                    "accountId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "transactionId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "account": {
                        "$ref": "#/components/schemas/Account"
                    },
                    "transaction": {
                        "$ref": "#/components/schemas/Transaction"
                    },
                    "source": {
                        "type": "boolean"
                    }
                },
                "additionalProperties": false
            },
...
            "Client": {
                "required": [
                    "email",
                    "name"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "maxLength": 100,
                        "minLength": 1,
                        "type": "string"
                    },
                    "email": {
                        "minLength": 1,
                        "type": "string",
                        "format": "email"
                    },
                    "accounts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Account"
                        },
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
            "ClientRequest": {
                "required": [
                    "email",
                    "name"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 100,
                        "minLength": 1,
                        "type": "string"
                    },
                    "email": {
                        "minLength": 1,
                        "type": "string",
                        "format": "email"
                    }
                },
                "additionalProperties": false
            },
...
            "ProblemDetails": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "detail": {
                        "type": "string",
                        "nullable": true
                    },
                    "instance": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": {}
            },
            "Rate": {
                "required": [
                    "currency"
                ],
                "type": "object",
                "properties": {
                    "currency": {
                        "maxLength": 3,
                        "minLength": 1,
                        "type": "string"
                    },
                    "currencyRate": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "additionalProperties": false
            },
...
            "Transaction": {
                "required": [
                    "currency"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "accountsTransactions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AccountTransaction"
                        },
                        "nullable": true
                    },
                    "amount": {
                        "type": "number",
                        "format": "double"
                    },
                    "currency": {
                        "maxLength": 3,
                        "minLength": 1,
                        "type": "string"
                    },
                    "rate": {
                        "$ref": "#/components/schemas/Rate"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "additionalProperties": false
            },
...
        }
    }
}